project/requirements.txt
asyncpg==0.22.0
black==20.8b1
fastapi==0.63.0
flake8===3.8.4
gunicorn==20.0.4
isort==5.7.0
pytest==6.2.2
pytest-cov==2.11.1
requests==2.25.1
tortoise-orm==0.16.21
uvicorn==0.13.4
